Create html-base
This is the full HTML for an Avatar Legends character sheet in Roll20. Combined with the translation and the CSS pages to get a fully working character sheet. This sheet is a heavily modified Monster of the Week sheet. I didn't invent any of this code, merely adapted and used it to create this resource. This sheet currently only has support the 6 playbooks in the Quickstart Guide, but it does have support for "Custom" playbooks and moves so that you can add other playbooks that you need.